+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     * Fetching badge data
     * Fetching due badges (added stock\_levels attribute, since OSM now gives it)
     * Fetching & updating badge links for Events
+    * Fetching & updating badge links for Activities
     * Summary now returns all started/completed/awarded badges (it can't filter by type) so can only be called from Osm::Badge
     * Osm::Badge
       * osm\_key, osm\_long\_key and competion criteria attributes are gone
@@ -23,6 +24,7 @@
       * Attributes added: badge\_name, badge\_id, badge\_version, requirement\_id
       * Attributes removed: badge\_label, badge\_key, requirement\_key
       * Attributes untouched: badge\_type, badge\_section, data, requirement\_label
+    * Osm::Activity::Badge attributes now match Osm::Event::BadgeLink
 
 
 ## Version 1.2.14